requirements.txt
# Core Deep Learning Framework
torch==2.1.0
torchvision==0.16.0
torchaudio==2.1.0

# Data Processing and Manipulation
numpy==1.24.3
pandas==2.0.3
opencv-python==4.8.1.78
pillow==10.0.1
scikit-learn==1.3.2

# Explainable AI Libraries
shap==0.44.0
lime==0.2.0.1
captum==0.7.0  # PyTorch library for model interpretability

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# Model Utilities
tqdm==4.66.1

# Image Augmentation
albumentations==1.3.1

# Additional Utilities
pyyaml==6.0.1
tensorboard==2.15.1